+++ capstone_flutter/lib/view/screen/wifi_screen/payment_methode_screen.dart
@@ -12,8 +12,8 @@ class PaymentMethodWifi extends StatefulWidget {
   final String pelangganData;
   final DateTime createdAt;
   final String providerName;
-  final int price;
-  final int adminFee;
+  final double price;
+  final double adminFee;
   final String customerName;
   const PaymentMethodWifi({
     super.key,
